🎨 投稿一覧ページの改善 #13
ロジックとビューの分離を図る。処理はusePostで定義し、ページの方で呼び出すようにする。
