+++ src/pages/index.tsx
@@ -1,13 +1,10 @@
 import type { NextPage } from "next";
-import Auth from "src/features/auth/Auth";
-import Layout from "src/components/Layout";
+import Auth from "src/features/auth/components/Auth";
+import Layout from "src/components/layouts/Layout";
+import { Auth2 } from "src/features/auth/components/Auth2";
 
 const Home: NextPage = () => {
-  return (
-    <Layout showHeader={false} showFooter={false}>
-      <Auth />
-    </Layout>
-  );
+  return <Auth2 />;
 };
 
 export default Home;
